Consolidate system cert healthchecks into single module
Merge expiration.py and trustflags.py into systemcerts.py to combine
trust flag validation and expiration checks into unified classes for
each subsystem (CA, KRA, OCSP, TKS, TPS) to reduce pki-healthcheck
execution time.
