+++ base/server/healthcheck/setup.py
@@ -32,8 +32,7 @@ setup(
         ],
         # plugin modules for pkihealthcheck.certs registry
         'pkihealthcheck.certs': [
-            'trust_flags = pki.server.healthcheck.certs.trustflags',
-            'expiration = pki.server.healthcheck.certs.expiration',
+            'systemcerts = pki.server.healthcheck.certs.systemcerts',
         ],
     },
     classifiers=[
